refactor: delete resource cache service shell
