--- backend/web/services/resource_cache.py
@@ -1,17 +1,0 @@
-"""Compatibility shell for monitor resource overview cache."""
-
-from backend.monitor.infrastructure.resources.resource_overview_cache import (
-    ResourceOverviewContractError,
-    clear_resource_overview_cache,
-    get_resource_overview_snapshot,
-    refresh_resource_overview_sync,
-    resource_overview_refresh_loop,
-)
-
-__all__ = [
-    "ResourceOverviewContractError",
-    "clear_resource_overview_cache",
-    "get_resource_overview_snapshot",
-    "refresh_resource_overview_sync",
-    "resource_overview_refresh_loop",
-]
