+++ tests/Unit/backend/web/services/test_agent_runtime_gateway_thread_input.py
@@ -59,7 +59,7 @@ async def test_gateway_thread_input_clears_resource_overview_cache_when_starting
         patch("backend.web.services.agent_pool.resolve_thread_sandbox", return_value="local"),
         patch("backend.web.services.agent_pool.get_or_create_agent", AsyncMock(return_value=agent)),
         patch("backend.web.services.streaming_service.start_agent_run", return_value="run-123"),
-        patch("backend.web.services.resource_cache.clear_resource_overview_cache") as clear_cache,
+        patch("backend.monitor.infrastructure.resources.resource_overview_cache.clear_resource_overview_cache") as clear_cache,
     ):
         result = await build_agent_runtime_gateway(app).dispatch_thread_input(_thread_input())
 
@@ -75,7 +75,7 @@ async def test_gateway_thread_input_requires_agent_runtime() -> None:
         patch("backend.web.services.agent_pool.resolve_thread_sandbox", return_value="local"),
         patch("backend.web.services.agent_pool.get_or_create_agent", AsyncMock(return_value=SimpleNamespace())),
         patch("backend.web.services.streaming_service.start_agent_run", return_value="run-123"),
-        patch("backend.web.services.resource_cache.clear_resource_overview_cache"),
+        patch("backend.monitor.infrastructure.resources.resource_overview_cache.clear_resource_overview_cache"),
     ):
         with pytest.raises(AttributeError):
             await build_agent_runtime_gateway(app).dispatch_thread_input(_thread_input())
@@ -90,7 +90,7 @@ async def test_gateway_thread_input_passes_enable_trajectory_to_start_agent_run(
         patch("backend.web.services.agent_pool.resolve_thread_sandbox", return_value="local"),
         patch("backend.web.services.agent_pool.get_or_create_agent", AsyncMock(return_value=agent)),
         patch("backend.web.services.streaming_service.start_agent_run", return_value="run-123") as start_run,
-        patch("backend.web.services.resource_cache.clear_resource_overview_cache"),
+        patch("backend.monitor.infrastructure.resources.resource_overview_cache.clear_resource_overview_cache"),
     ):
         await build_agent_runtime_gateway(app).dispatch_thread_input(_thread_input(enable_trajectory=True))
